Prioritize downloaded spdlog headers
In case you have spdlog installed on the build machine, and the system
SoapySDR is being used, it will try to use the system spdlog headers
instead of the headers from the downloaded spdlog. This will cause a
conflict and linking errors. To avoid this, prioritize the downloaded
spdlog headers.